requirements.txt
PyQt5
python-barcode
PyQtWebEngine
pyinstaller